Throw error when setting array element of bad size
The `array` implementation for arbitrary serialisable objects expects
the object type to be of fixed size after serialization.  This size is
determined by creating an object using the default constructor and then
serializing it.
If later on an object is added that serialises to a different number of
bytes (e.g. because it contains dynamically allocated members that
differ in size compared to the test-object created by the default
constructor), this results in an obscure cereal error later on.

To provide a more understandable error message to the user, check the
size of each element when adding it to the array and throw a
runtime_error if it does not match the expected size.